Fix #59: End the study if users agree to sign up for sync.
Also includes some fixes for previously-existing ways for ending
the study. The Advisor class is now an EventEmitter (copied from
the SHIELD utils lib because it's not exported) and emits events
when the user agrees to the study, asks to disable it completely,
or signs up for sync while the study is running.

Because the SHIELD utils don't currently have a good way for ending
a study with a custom reason, we manually log the final telemetry
packet, trigger the survey, and end the study.